requirements.txt
certifi==2022.12.7
emoji==2.2.0
future==0.18.3
python-telegram-bot==10.1.0
